🐛 Fix file type js to jsx
